app/requirements.txt
flask
flask-cors
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
vertexai==1.71.1
